Put dash at the end of character class in example i18n.js code. (#797)
* Put dash at the end of character class in example i18n.js code.

The regexp /[0-9-_]/ matches digits, hyphen, and underscore, but only
because the hyphen comes right after a hyphenated range.  If you put
anything in between the 9 and the hyphen, boom.  It's so surprising, I
think it's clearer just to put the hyphen at the end.

* Anchor regexp to beginning of filename.

This regular expression is trying to find files of the form "foo.json",
where "foo" doesn't contain any special characters.  But the regex is
used as a substring match, so it needs to anchor to the beginning of the
basename, otherwise it will match anything which ends with "x.json".

Co-authored-by: Justin Lebar <jlebar@google.com>