Fix flow types and enable flow types testing on CI (#896)
* Fix broken unit tests

The tests were using "assert" function wrong.
"assert" is defined as: assert(value, [message])

The tests passed two arguments in an attempt to compare them while
the only thing that happened was that the first value was checked for
being truthy.

* Fix flow types and enable flow testing on CI

Types fixed by making "getChoiceIndex" method follow types and be
a class property rather than class method. Class methods are "read-only"
in flow and that wasn't compatible with flow types that are defining
that function as a property (to allow overriding through prototype changing).